feat: adjust book module with find by user id, and vinculate user in creation

+++ server/src/modules/book/book.service.ts
@@ -1,19 +1,27 @@
-import { Injectable } from '@nestjs/common';
+import { HttpCode, HttpException, HttpStatus, Inject, Injectable, Scope } from '@nestjs/common';
 import { CreateBookDto } from './dto/create-book.dto';
 import { FindAllDto } from './dto/find-all.dto';
 import { BookRepository } from '../repositories/bookRepository.service';
+import { REQUEST } from '@nestjs/core';
+import { IUserToken } from '@/interfaces/IUserToken.interface';
 
-@Injectable()
+@Injectable({ scope: Scope.REQUEST })
 export class BookService {
-  constructor(private readonly bookRepository: BookRepository) {}
-  create(data: CreateBookDto) {
-    return this.bookRepository.create(data);
+  constructor(
+    private readonly bookRepository: BookRepository,
+    @Inject(REQUEST)
+    private readonly request: Request & IUserToken,
+  ) {}
+  async create(data: CreateBookDto) {
+    const book = await this.bookRepository.findOne(data.bookId);
+    if (book) throw new HttpException('Book already exists', HttpStatus.CONFLICT);
+    return this.bookRepository.create(data, this.request);
   }
 
   async findAll(query: FindAllDto) {
     return {
-      books: await this.bookRepository.findAll(query),
-      total: await this.bookRepository.countAll(),
+      books: await this.bookRepository.findAll(query, this.request),
+      total: await this.bookRepository.countAll(this.request),
     };
   }
 
@@ -21,7 +29,8 @@ export class BookService {
     return this.bookRepository.findWithoutFilters();
   }
 
-  remove(id: string) {
-    return this.bookRepository.delete(id);
+  async remove(id: string) {
+    const book = await this.bookRepository.findOne(id);
+    if (book) return this.bookRepository.delete(id);
   }
 }
